Revert "Initial changes for olugin mgr"
This reverts commit b98ec04c2e2786063879699103d8e252c11c0a52.
