+++ go/src/lib/lib_test/tx_test.go
@@ -22,8 +22,8 @@ import (
     "io"
     "log/syslog"
     "net/rpc"
-    . "gocode/src/lib/lomcommon"
-    . "gocode/src/lib/lomipc"
+    . "lib/lomcommon"
+    . "lib/lomipc"
     "os"
     "strconv"
     "testing"
