+++ go/src/lib/lomipc/server_transport.go
@@ -1,13 +1,13 @@
 package lomipc
 
 import (
-	"encoding/gob"
-	"fmt"
-	. "gocode/src/lib/lomcommon"
-	"net"
-	"net/http"
-	"net/rpc"
-	"reflect"
+    "encoding/gob"
+    "fmt"
+    . "lib/lomcommon"
+    "net"
+    "net/http"
+    "net/rpc"
+    "reflect"
 )
 
 /*
@@ -16,11 +16,11 @@ import (
  *  NOTE:
  *      This is used as only IPC between processes running inside a single
  *      container as single system; Tightly coupled with static set of APIs
- *
+ *      
  *  Multiple clients (PluginMgr) send requests concurrently to a server/engine.
  *  Server process each sequentially/concurrently and respond back to each as succeeded
  *  or failed.
- *
+ *  
  *  Server send requests to multiple clients sequentially/concurrently as a request
  *  addressed to a client only. A single client may receive multiple requests at any time.
  *  For each received request, client confirm back as succeeded / failed, synchronously.
@@ -37,9 +37,9 @@ import (
  *  plugins managed by client.
  */
 
+
 /* All types of requests from client to server */
 type ReqDataType int
-
 const (
     TypeNone = ReqDataType(iota)
     TypeRegClient
@@ -52,37 +52,37 @@ const (
     TypeCount
 )
 
-var ReqTypeToStr = map[ReqDataType]string{
-	TypeNone:                  "None",
-	TypeRegClient:             "RegisterClient",
-	TypeDeregClient:           "DeregisterClient",
-	TypeRegAction:             "RegisterAction",
-	TypeDeregAction:           "DeregisterAction",
-	TypeRecvServerRequest:     "RecvServerRequest",
-	TypeSendServerResponse:    "SendServerResponse",
-	TypeNotifyActionHeartbeat: "NotifyActionHeartbeat",
+var ReqTypeToStr = map[ReqDataType]string {
+    TypeNone: "None",
+    TypeRegClient: "RegisterClient",
+    TypeDeregClient: "DeregisterClient",
+    TypeRegAction: "RegisterAction",
+    TypeDeregAction: "DeregisterAction",
+    TypeRecvServerRequest: "RecvServerRequest",
+    TypeSendServerResponse: "SendServerResponse",
+    TypeNotifyActionHeartbeat: "NotifyActionHeartbeat",
 }
 
-/* Server sends its request as response to TypeRecvServerRequest */
+/* Server sends its request as response to TypeRecvServerRequest */ 
 type ServerReqDataType int
-
 const (
-	TypeServerRequestAction = ServerReqDataType(iota)
-	TypeServerRequestShutdown
-	TypeServerRequestCount
+    TypeServerRequestAction = ServerReqDataType(iota)
+    TypeServerRequestShutdown
+    TypeServerRequestCount
 )
 
-var ServerReqTypeToStr = map[ServerReqDataType]string{
-	TypeServerRequestAction:   "RecvServerRequestAction",
-	TypeServerRequestShutdown: "RecvServerRequestShutdown",
+
+var ServerReqTypeToStr = map[ServerReqDataType]string {
+    TypeServerRequestAction: "RecvServerRequestAction",
+    TypeServerRequestShutdown: "RecvServerRequestShutdown",
 }
 
 /* Request from client to server over RPC */
 type LoMRequest struct {
-	ReqType     ReqDataType /* Type of request */
-	Client      string      /* The client sending this request */
-	TimeoutSecs int         /* Timeout - Honored in long running requests */
-	ReqData     interface{} /* Data specific to request type */
+    ReqType     ReqDataType     /* Type of request */
+    Client      string          /* The client sending this request */
+    TimeoutSecs int             /* Timeout - Honored in long running requests */
+    ReqData     interface{}     /* Data specific to request type */
 }
 
 /*
@@ -94,41 +94,41 @@ type LoMRequest struct {
  * RespData is specific to client's request. It could be nil.
  */
 type LoMResponse struct {
-	ResultCode int
-	ResultStr  string
+    ResultCode  int
+    ResultStr   string
 
-	RespData interface{}
+    RespData    interface{}
 }
 
 /*
  * Msg to pass via ReqData in LomRequest
  * All kinds of msg data matching request type
  */
-type MsgRegClient struct { /* For TypeRegClient */
+type MsgRegClient struct {          /* For TypeRegClient */
 }
 
-type MsgDeregClient struct { /* For TypeDeregClient */
+type MsgDeregClient struct {        /* For TypeDeregClient */
 }
 
-type MsgRegAction struct { /* For TypeRegAction */
-	Action string
+type MsgRegAction struct {          /* For TypeRegAction */
+    Action  string
 }
 
-type MsgDeregAction struct { /* For TypeDeregAction */
-	Action string
+type MsgDeregAction struct {        /* For TypeDeregAction */
+    Action  string
 }
 
-type MsgRecvServerRequest struct { /* For TypeRecvServerRequest */
+type MsgRecvServerRequest struct {  /* For TypeRecvServerRequest */
 }
 
 type MsgSendServerResponse struct { /* For TypeSendServerResponse */
-	ReqType ServerReqDataType
-	ResData interface{} /* ActionResponseData */
+    ReqType             ServerReqDataType
+    ResData             interface {}
 }
 
-type MsgNotifyHeartbeat struct { /* For TypeNotifyActionHeartbeat */
-	Action    string
-	Timestamp int64
+type MsgNotifyHeartbeat struct {    /* For TypeNotifyActionHeartbeat */
+    Action      string
+    Timestamp   int64
 }
 
 /*
@@ -139,9 +139,9 @@ type MsgNotifyHeartbeat struct { /* For TypeNotifyActionHeartbeat */
  * The ReqData is specific per type.
  */
 type ServerRequestData struct {
-	ReqType ServerReqDataType /* Type of requests from server to client */
-	ReqData interface{}       /* Data per request type */
-	/* ActionRequestData or ShutdownRequestData */
+    ReqType             ServerReqDataType   /* Type of requests from server to client */
+    ReqData             interface {}        /* Data per request type */
+                                            /* ActionRequestData or ShutdownRequestData */
 }
 
 /*
@@ -149,34 +149,34 @@ type ServerRequestData struct {
  * MsgSendServerResponse::ReqType == TypeServerRequestAction
  */
 type ActionResponseData struct {
-	Action            string
-	InstanceId        string
-	AnomalyInstanceId string
-	AnomalyKey        string
-	Response          string
-	ResultCode        int
-	ResultStr         string
+    Action              string
+    InstanceId          string
+    AnomalyInstanceId   string
+    AnomalyKey          string
+    Response            string
+    ResultCode          int
+    ResultStr           string
 }
 
 /* Helper to convert ActionResponseData as Map */
 func (p *ActionResponseData) ToMap(end bool) map[string]string {
-	ret := map[string]string{
-		"action":            p.Action,
-		"instanceId":        p.InstanceId,
-		"anomalyInstanceId": p.AnomalyInstanceId,
-		"anomalyKey":        p.AnomalyKey,
-		"response":          p.Response,
-		"resultCode":        fmt.Sprintf("%d", p.ResultCode),
-		"resultStr":         p.ResultStr,
-	}
-	if p.InstanceId == p.AnomalyInstanceId {
-		if end {
-			ret["state"] = "complete"
-		} else {
-			ret["state"] = "init"
-		}
-	}
-	return ret
+    ret := map[string]string {
+        "action": p.Action,
+        "instanceId": p.InstanceId,
+        "anomalyInstanceId": p.AnomalyInstanceId,
+        "anomalyKey": p.AnomalyKey,
+        "response": p.Response,
+        "resultCode": fmt.Sprintf("%d", p.ResultCode),
+        "resultStr": p.ResultStr,
+    }
+    if p.InstanceId == p.AnomalyInstanceId {
+        if end {
+            ret["state"] = "complete"
+        } else {
+            ret["state"] = "init"
+        }
+    }
+    return ret
 }
 
 /* Helper to validate ActionResponseData */
@@ -194,17 +194,18 @@ func (p *ActionResponseData) Validate() bool {
     return true
 }
 
+
 /*
  * ReqData for ServerRequestData::ReqData for
  * ServerRequestData::ReqType == TypeServerRequestAction
  */
 type ActionRequestData struct {
-	Action            string
-	InstanceId        string
-	AnomalyInstanceId string
-	AnomalyKey        string
-	Timeout           int
-	Context           []*ActionResponseData
+    Action              string
+    InstanceId          string
+    AnomalyInstanceId   string
+    AnomalyKey          string
+    Timeout             int
+    Context             []*ActionResponseData
 }
 
 /*
@@ -235,14 +236,14 @@ func SlicesComp(p []*ActionResponseData, q []*ActionResponseData) bool {
 
 /* Helper to compare given requests. */
 func (r *ActionRequestData) Equal(p *ActionRequestData) bool {
-	if r == p {
-		/* Same ptr */
-		return true
-	}
-	if (r == nil) || (p == nil) {
-		LogError("Unexpected nil args self(%v) arg(%v)\n", (r == nil), (p == nil))
-		return false
-	}
+    if r == p {
+        /* Same ptr */
+        return true
+    }
+    if (r == nil) || (p == nil) {
+        LogError("Unexpected nil args self(%v) arg(%v)\n", (r == nil), (p == nil))
+        return false
+    }
 
     if ((r.Action == p.Action) &&
         (r.InstanceId == p.InstanceId) &&
@@ -254,6 +255,7 @@ func (r *ActionRequestData) Equal(p *ActionRequestData) bool {
         return false
     }
 }
+
 
 /* Helper to compare given requests. */
 func (r *ServerRequestData) Equal(p *ServerRequestData) bool {
@@ -294,87 +296,89 @@ func (r *ServerRequestData) Equal(p *ServerRequestData) bool {
     }
 }
 
+
 /*
  * Each proc has a channel for remote end to write request.
  * Each request carry a channel for response to that request.
  */
 type LoMTransport struct {
-	ServerCh chan interface{}
+    ServerCh    chan interface{}
 }
 
 /* Internal req object within server. */
 type LoMRequestInt struct {
-	Req *LoMRequest
-	/* LoMResponse to this request is sent via this chan */
-	ChResponse chan interface{}
-}
+    Req         *LoMRequest
+    /* LoMResponse to this request is sent via this chan */
+    ChResponse  chan interface{}
+} 
 
 /* RPC call from client */
 func (tr *LoMTransport) SendToServer(req *LoMRequest, reply *LoMResponse) (err error) {
 
-	defer func() {
-		if err != nil {
-			LogError("SendToServer cl(%s) mtype(%s) failed (%v)",
-				req.Client, ReqTypeToStr[req.ReqType], err)
-		} else {
-			LogInfo("SUCCESS: SendToServer cl(%s) mtype(%s) result(%d)/(%s)", req.Client,
-				ReqTypeToStr[req.ReqType], reply.ResultCode, reply.ResultStr)
-		}
-	}()
+    defer func() {
+        if err != nil {
+            LogError("SendToServer cl(%s) mtype(%s) failed (%v)", 
+                    req.Client, ReqTypeToStr[req.ReqType], err)
+        } else {
+            LogInfo("SUCCESS: SendToServer cl(%s) mtype(%s) result(%d)/(%s)", req.Client,
+                    ReqTypeToStr[req.ReqType], reply.ResultCode, reply.ResultStr)
+        }
+    } ()
 
-	if (req == nil) || (reply == nil) {
-		return LogError("Nil args req(%v) reply(%v)", req, reply)
-	}
+    if (req == nil) || (reply == nil) {
+        return LogError("Nil args req(%v) reply(%v)", req, reply)
+    }
 
-	rpcReq := LoMRequestInt{req, make(chan interface{}, 1)}
-	tr.ServerCh <- &rpcReq
+    rpcReq := LoMRequestInt { req, make(chan interface{}, 1) }
+    tr.ServerCh <- &rpcReq
 
-	LogDebug("Req sent to server client(%s) type(%s). Waiting for response...",
-		req.Client, ReqTypeToStr[req.ReqType])
+    LogDebug("Req sent to server client(%s) type(%s). Waiting for response...",
+            req.Client, ReqTypeToStr[req.ReqType])
 
-	/* Wait for server response */
-	p := <-rpcReq.ChResponse
-	if x, ok := p.(*LoMResponse); ok {
-		*reply = *x
-	} else {
-		return LogError("Server response message (%T) != *LoMResponse", x)
-	}
+    /* Wait for server response */
+    p := <- rpcReq.ChResponse
+    if x, ok := p.(*LoMResponse); ok {
+        *reply = *x
+    } else {
+        return LogError("Server response message (%T) != *LoMResponse", x)
+    }
 
-	return nil
+    return nil
 }
 
 /* Local call from server to read client request. */
 func (tr *LoMTransport) ReadClientRequest(chAbort chan interface{}) (*LoMRequestInt, error) {
-	/* Return on non-null request or upon abort */
-	select {
-	case p := <-tr.ServerCh:
-		if x, ok := p.(*LoMRequestInt); ok {
-			LogDebug("Server: Read from client (%s) type(%s)", x.Req.Client, ReqTypeToStr[x.Req.ReqType])
-			return x, nil
-			/* Let server return response upon processing, via channel embedded in msg. */
-		} else {
-			return nil, LogError("Client request message (%T) != *Msg", x)
-		}
-
-	case <-chAbort:
-		return nil, LogError("Server: Aborting read via abort channel")
-		/* Aborting per instruction */
-	}
+    /* Return on non-null request or upon abort */
+    select {
+    case p := <-tr.ServerCh:
+        if x, ok := p.(*LoMRequestInt); ok {
+            LogDebug("Server: Read from client (%s) type(%s)", x.Req.Client, ReqTypeToStr[x.Req.ReqType])
+            return x, nil
+            /* Let server return response upon processing, via channel embedded in msg. */
+        } else {
+            return nil, LogError("Client request message (%T) != *Msg", x)
+        }
+        
+    case <- chAbort:
+        return nil, LogError("Server: Aborting read via abort channel")
+        /* Aborting per instruction */
+    }
 }
 
+
 func init_encoding() {
-	gob.Register(MsgRegClient{})
-	gob.Register(MsgDeregClient{})
-	gob.Register(MsgRegAction{})
-	gob.Register(MsgDeregAction{})
-	gob.Register(MsgRecvServerRequest{})
-	gob.Register(MsgSendServerResponse{})
-	gob.Register(MsgNotifyHeartbeat{})
-	gob.Register(ServerRequestData{})
-	gob.Register(ActionRequestData{})
-	gob.Register(ActionResponseData{})
-	gob.Register(ShutdownRequestData{})
-	gob.Register(MsgEmptyResp{})
+    gob.Register(MsgRegClient{})
+    gob.Register(MsgDeregClient{})
+    gob.Register(MsgRegAction{})
+    gob.Register(MsgDeregAction{})
+    gob.Register(MsgRecvServerRequest{})
+    gob.Register(MsgSendServerResponse{})
+    gob.Register(MsgNotifyHeartbeat{})
+    gob.Register(ServerRequestData{})
+    gob.Register(ActionRequestData{})
+    gob.Register(ActionResponseData{})
+    gob.Register(ShutdownRequestData{})
+    gob.Register(MsgEmptyResp{})
 }
 
 func GetLoMTransport() *LoMTransport {
@@ -389,14 +393,17 @@ func GetLoMTransport() *LoMTransport {
 func ServerInit() (*LoMTransport, error) {
     tr := GetLoMTransport()
 
-	rpc.Register(tr)
-	rpc.HandleHTTP()
-	l, e := net.Listen("tcp", ":1234")
-	if e != nil {
-		LogPanic("listen error:(%v)", e)
-		return nil, e
-	}
-	go http.Serve(l, nil)
-	LogDebug("Server: Started serving")
-	return tr, nil
+    rpc.Register(tr)
+    rpc.HandleHTTP()
+    l, e := net.Listen("tcp", ":1234")
+    if e != nil {
+        LogPanic("listen error:(%v)", e)
+        return nil, e
+    }
+    go http.Serve(l, nil)
+    LogDebug("Server: Started serving")
+    return tr, nil
 }
+
+
+
