--- gocode/src/pluginmgr/pluginmgr_main.go
@@ -1,23 +1,0 @@
-package main
-
-import (
-	"gocode/src/lib/lomcommon"
-	"gocode/src/pluginmgr/pluginmgr_common"
-	"log/syslog"
-)
-
-func main() {
-	lomcommon.SetLogLevel(syslog.LOG_NOTICE)
-	lomcommon.LogMessage(syslog.LOG_NOTICE, "plugin_mgr : Starting Plugin Manager")
-
-	if err := pluginmgr_common.SetupPluginManager(); err != nil {
-		lomcommon.LogMessage(syslog.LOG_ERR, "plugin_mgr : SetupPluginManager failed")
-		return
-	}
-	if err := pluginmgr_common.StartPluginManager(); err != nil {
-		lomcommon.LogMessage(syslog.LOG_ERR, "plugin_mgr : StartPluginManager failed")
-		return
-	}
-
-	lomcommon.LogMessage(syslog.LOG_NOTICE, "plugin_mgr : Exiting plugin manager")
-}
