--- gocode/src/pluginmgr/pluginmgr_test/globals_conf.json
@@ -1,5 +1,0 @@
-{
-	"MAX_SEQ_TIMEOUT_SECS" : 120,
-    "MIN_PERIODIC_LOG_PERIOD_SECS" : 15,
-    "ENGINE_HB_INTERVAL_SECS" : 10
-}
