--- gocode/src/pluginmgr/pluginmgr_test/proc_conf.json
@@ -1,21 +1,0 @@
-{
-	"proc_0": {
-		"link_crc": {
-			"name": "link_crc",
-			"version": "00.01.1",
-			"path": " /path/"
-		},
-		"link_flap": {
-			"name": "link_flap",
-			"version": "02.00.1",
-			"path": " /path/"
-		}
-	},
-	"proc_1": {
-		"bgp_holddown": {
-			"name": "bgp_holddown",
-			"version": "02_1",
-			"path": " /path/"
-		}
-	}
-}
